requirements.txt
arxiv
google-genai
dotenv
os
streamlit